Refactor structures
