+++ src/api/lib/infrastructure/adapter/openai/OpenAiService.ts
