+++ src/api/service/translation/TranslationService.test.ts
@@ -1,8 +1,8 @@
 import { Language, LanguagesType } from "@/api/lib/domain/translator/Language";
 import { OriginalText } from "@/api/lib/domain/translator/OriginalText";
 import { TranslationConfig } from "@/api/lib/domain/translator/TranslationConfig";
-import { TranslationServiceImpl } from "@/api/service/adapter/TranslationService";
-import { OpenAiService } from "@/api/service/openai/OpenAiService";
+import { TranslationServiceImpl } from "@/api/service/translation/TranslationService";
+import { OpenAiService } from "@/api/lib/infrastructure/adapter/openai/OpenAiService";
 
 describe("TranslationServiceImpl", () => {
   it("should translate text correctly", async () => {
